Determine polarity of review given by user
It is a sentiment analysis problem used to classify the polarity of a review given by amazon users.
Given a review, we will determine whether it is positive or negative. We use the ratings to predict the kind of review. A rating of 4 or 5 is considered a positive review. A review of 1 or 2 is considered as a negative review. A review of 3 is considered to be neutral and hence ignored.